APIHandler() & Repository
Created an APIHandler class which will be called in all repositories. Also renamed BudgetRepository to Repository as this file will be used as the generic repository for all functions in the project.
